refactor: decompose ModuleDetailPage into hook + 5 sub-components
Extract useModuleDetail hook (all data fetching + state) and 5 panel
components: ModuleDocumentation, SecurityScanPanel, SCMRepositoryPanel,
WebhookEventsPanel, VersionDetailsPanel. Reduces ModuleDetailPage from
1,133 to 444 lines.

Closes #102
