+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- refactor: decompose `ModuleDetailPage.tsx` (1,133 → 444 lines) into `useModuleDetail` hook and 5 sub-components — `ModuleDocumentation`, `SecurityScanPanel`, `SCMRepositoryPanel`, `WebhookEventsPanel`, `VersionDetailsPanel`
+
 ## [0.4.0] - 2026-04-13
 
 ### Added
